requirements.txt
pillow==10.1.0
numpy==1.26.2
scipy==1.11.4
moviepy==1.0.3
wave==0.0.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
schedule==1.2.0

